requirements.txt
# EconoScope Complete - All Phases
# Backend & Data
fastapi==0.109.0
uvicorn[standard]==0.25.0
pydantic==2.5.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
httpx==0.26.0

# Data Collection
beautifulsoup4==4.12.2
pandas==2.1.4
numpy==1.26.2
boto3==1.34.10

# ML
xgboost==2.0.3
scikit-learn==1.3.2
joblib==1.3.2

# SMS & Alerts
africastalking==1.2.8
celery[redis]==5.3.4
redis==5.0.1

# Cloud
gradient==2.0.6

# Dev
pytest==7.4.3
